requirements.txt
aniso8601==8.1.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.0.0
pytz==2020.4
requests==2.25.1
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.22
termcolor==1.1.0
urllib3==1.26.2
Werkzeug==1.0.1
